// apps/ui/shaddercn/components/dropdown/types.ts
export type Direction = "top" | "bottom" | "left" | "right";

// Anchor point alignment
export type Anchor = "start" | "center" | "end";